New Design for file overview
