+++ src/modules/file-explorer/routes.jsx
@@ -1,0 +1,9 @@
+import FileExplorer from './containers/file-explorer';
+
+export default [
+	{
+		path: '/files/',
+		name: 'file-explorer',
+		component: FileExplorer
+	}
+];
