+++ src/modules/file-explorer/styles/directories.scss
@@ -1,0 +1,27 @@
+@import '../../core/styles/colors.scss';
+
+.directories {
+	width:100%;
+	float:left;
+	margin-bottom:40px;
+
+	h5 {
+		color: $colorHPIRed;
+		padding:10px 0px 20px;
+	}
+
+	.card.folder {
+		box-shadow: 0px 1px 2px rgba(0,0,0,0.1);
+
+		&:hover {
+			box-shadow: 0px 1px 4px rgba(0,0,0,0.2);
+			background:#f9f9f9;
+			cursor: pointer;
+		}
+
+		.fa {
+			margin-right:15px;
+			color:#ccc;
+		}
+	}
+}
