+++ src/modules/file-explorer/styles/file-explorer.scss
@@ -1,5 +1,5 @@
 @import '../../core/styles/colors.scss';
 
-.route-fileExplorer {
+.route-file-explorer {
 
 }
