+++ src/modules/file-explorer/styles/upload.scss
@@ -1,0 +1,19 @@
+@import '../../core/styles/colors.scss';
+
+.section-upload {
+	width:100%;
+	float:left;
+	padding:40px 0px;
+
+	.drop-zone {
+		width:100%;
+		height:200px;
+		border:4px dashed #ddd;
+		background:#f9f9f9;
+		float:left;
+		font-size:30px;
+		text-align:center;
+		color:#ccc;
+		line-height:192px;
+	}
+}
