--- src/modules/fileExplorer/routes.jsx
@@ -1,9 +1,0 @@
-import FileExplorer from './containers/fileExplorer';
-
-export default [
-	{
-		path: '/files/',
-		name: 'fileExplorer',
-		component: FileExplorer
-	}
-];
