--- src/modules/fileExplorer/styles/directories.scss
@@ -1,8 +1,0 @@
-@import '../../core/styles/colors.scss';
-
-.directories {
-  border-bottom:2px solid #696969;
-  .card {
-    margin-right: 10px;
-  }
-}
