--- src/modules/fileExplorer/styles/files.scss
@@ -1,12 +1,0 @@
-@import '../../core/styles/colors.scss';
-
-.files {
-  padding-top: 5%;
-  .card {
-    margin-right: 10px;
-  }
-  .card-img-top {
-		max-width:100%;
-		height:auto;
-	}
-}
