--- src/modules/fileExplorer/styles/memory.scss
@@ -1,16 +1,0 @@
-@import '../../core/styles/colors.scss';
-
-.memory {
-  padding-top: 4%;
-
-  .rightSide {
-    float: right;
-    text-align: right;
-  }
-}
-
-.upload {
-  span {
-    padding-left: 5px;
-  }
-}
